pyproject.toml
[tool.poetry]
name = "pp"
version = "0.1.0"
description = ""
authors = ["Vitalii Yarmus"]

[tool.poetry.dependencies]
python = "3.7.7"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"
alembic = "^1.4.3"
Flask-SQLAlchemy = "^2.4.4"
Flask-Script = "^2.0.6"
Flask-Migrate = "^2.5.3"
SQLAlchemy-Utils = "^0.36.8"
PyMySQL = "^0.10.1"
cryptography = "^3.3.1"
Flask-Inputs = "^0.3.0"
jsonschema = "^3.2.0"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = 'pp:start'
